First fully functioning version.
Added APOC command push functionality.

Next step:
- Make sure the program plays nicely.
Setup a list of tests to benchmark the program. (Design appropriat
failure)
- Add logging and mail notifications.
- Add appropriate program paramters.
- Add catch for outerloop, close UDP socket and email admin.
- Update web application client with new URI's.


Realization: 
This version of the NIP requires a public IP address in order for the
NSCL to make an APOC push. Whereas the previous branch does not. This is
not a problem since the NIP will run on the same machine as the NSCL (in
the cloud). But it is something worth keeping in mind for future
designs.